+++ docs/conf.py
@@ -26,7 +26,7 @@ copyright = "2019, Open Contracting Partnership"
 author = "Open Contracting Partnership"
 
 # The short X.Y version
-version = "0.6.6"
+version = "0.7.0"
 # The full version, including alpha/beta/rc tags
 release = version