+++ docs/changelog.rst
@@ -8,7 +8,7 @@ Changed
 ~~~~~~~
 
 -  :func:`ocdsmerge.util.get_tags` raises an error if it encounters an HTTP error.
--  Drop support for Python 3.6 (end-of-life 2021-12-23).
+-  Drop support for Python 3.6 and 3.7.
 
 Added
 ~~~~~